+++ MatomoTracker/EventAPISerializer.swift
@@ -94,6 +94,7 @@ fileprivate extension Event {
                 URLQueryItem(name: "ec_sh", value: orderShippingCost != nil ? "\(orderShippingCost!)" : nil),
                 URLQueryItem(name: "ec_dt", value: orderDiscount != nil ? "\(orderDiscount!)" : nil),
                 URLQueryItem(name: "_ects", value: lastOrderTimestamp),
+                URLQueryItem(name: "link", value: outlink?.absoluteString),
                 ].filter { $0.value != nil }
 
             let dimensionItems = dimensions.map { URLQueryItem(name: "dimension\($0.index)", value: $0.value) }